Adding helper methods for converting and filtering a list of dates.
Adding helper methods for to get closest previous or next date to today from an attribute in the context.

Adding helper methods to identify a word of a list of words is present in a text (not part of another word). This could be used to identify if the text is using a past tense verbs.

Removed to_date_list because it can already be done via intent_handlerdecorator:

@intent_handler('Whatever')
def handle(dates_list: List[datetime.date]):
    ...

Removed _str_remove_punctuation because it can already be done via l10n.nl_strip

moved close_next_date and closed_previous_date from intents to entities and make it independet from context.now by introducing a parameter for the date.

moved is_text_including_words from intents to entities